Move types for the rest of library interactives to separate files.
