+++ src/open-response/components/types.ts
@@ -5,8 +5,13 @@ import { IAuthoringOpenResponseMetadata, IRuntimeOpenResponseMetadata } from "@c
 
 export interface IAuthoredState extends IAuthoringOpenResponseMetadata {
   version: number;
+  questionType: any;
+  prompt: string;
   defaultAnswer?: string;
   hint?: string;
+  required: boolean;
+  predictionFeedback?: string;
+  customFeedback?: string;
 }
 
 export interface IInteractiveState extends IRuntimeOpenResponseMetadata {}
