+++ src/video-player/components/app.tsx
@@ -2,30 +2,8 @@ import React from "react";
 import { JSONSchema6 } from "json-schema";
 import { BaseQuestionApp } from "../../shared/components/base-question-app";
 import { Runtime } from "./runtime";
+import { IAuthoredState, IInteractiveState } from "./types";
 import { IAuthoringInteractiveMetadata, IRuntimeInteractiveMetadata } from "@concord-consortium/lara-interactive-api";
-
-// Note that TS interfaces should match JSON schema. Currently there's no way to generate one from the other.
-// TS interfaces are not available in runtime in contrast to JSON schema.
-
-export interface IAuthoredState extends IAuthoringInteractiveMetadata {
-  version: number;
-  videoUrl?: string;
-  captionUrl?: string;
-  poster?: string;
-  prompt?: string;
-  credit?: string;
-  creditLink?: string;
-  creditLinkDisplayText?: string;
-  fixedAspectRatio?: string;
-  fixedHeight?: number;
-  required?: boolean;
-}
-
-export interface IInteractiveState extends IRuntimeInteractiveMetadata {
-  percentageViewed: number;
-  lastViewedTimestamp: number;
-  submitted?: boolean;
-}
 
 export const baseAuthoringProps = {
   schema: {
